feat: implement gRPC server for the Variant RPC method
