+++ internal/server/evaluation/evaluator_test.go
@@ -1,0 +1,736 @@
+package evaluation
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	errs "go.flipt.io/flipt/errors"
+	"go.flipt.io/flipt/internal/storage"
+)
+
+func Test_matchesString(t *testing.T) {
+	tests := []struct {
+		name       string
+		constraint storage.EvaluationConstraint
+		value      string
+		wantMatch  bool
+	}{
+		{
+			name: "eq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "bar",
+			},
+			value:     "bar",
+			wantMatch: true,
+		},
+		{
+			name: "negative eq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "bar",
+			},
+			value: "baz",
+		},
+		{
+			name: "neq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "neq",
+				Value:    "bar",
+			},
+			value:     "baz",
+			wantMatch: true,
+		},
+		{
+			name: "negative neq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "neq",
+				Value:    "bar",
+			},
+			value: "bar",
+		},
+		{
+			name: "empty",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "empty",
+			},
+			value:     " ",
+			wantMatch: true,
+		},
+		{
+			name: "negative empty",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "empty",
+			},
+			value: "bar",
+		},
+		{
+			name: "not empty",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "notempty",
+			},
+			value:     "bar",
+			wantMatch: true,
+		},
+		{
+			name: "negative not empty",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "notempty",
+			},
+			value: "",
+		},
+		{
+			name: "unknown operator",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "foo",
+				Value:    "bar",
+			},
+			value: "bar",
+		},
+		{
+			name: "prefix",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "prefix",
+				Value:    "ba",
+			},
+			value:     "bar",
+			wantMatch: true,
+		},
+		{
+			name: "negative prefix",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "prefix",
+				Value:    "bar",
+			},
+			value: "nope",
+		},
+		{
+			name: "suffix",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "suffix",
+				Value:    "ar",
+			},
+			value:     "bar",
+			wantMatch: true,
+		},
+		{
+			name: "negative suffix",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "suffix",
+				Value:    "bar",
+			},
+			value: "nope",
+		},
+	}
+	for _, tt := range tests {
+		var (
+			constraint = tt.constraint
+			value      = tt.value
+			wantMatch  = tt.wantMatch
+		)
+
+		t.Run(tt.name, func(t *testing.T) {
+			match := matchesString(constraint, value)
+			assert.Equal(t, wantMatch, match)
+		})
+	}
+}
+
+func Test_matchesNumber(t *testing.T) {
+	tests := []struct {
+		name       string
+		constraint storage.EvaluationConstraint
+		value      string
+		wantMatch  bool
+		wantErr    bool
+	}{
+		{
+			name: "present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "present",
+			},
+			value:     "1",
+			wantMatch: true,
+		},
+		{
+			name: "negative present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "present",
+			},
+		},
+		{
+			name: "not present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "notpresent",
+			},
+			wantMatch: true,
+		},
+		{
+			name: "negative notpresent",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "notpresent",
+			},
+			value: "1",
+		},
+		{
+			name: "NAN constraint value",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "bar",
+			},
+			value:   "5",
+			wantErr: true,
+		},
+		{
+			name: "NAN context value",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "5",
+			},
+			value:   "foo",
+			wantErr: true,
+		},
+		{
+			name: "eq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "42.0",
+			},
+			value:     "42.0",
+			wantMatch: true,
+		},
+		{
+			name: "negative eq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "42.0",
+			},
+			value: "50",
+		},
+		{
+			name: "neq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "neq",
+				Value:    "42.0",
+			},
+			value:     "50",
+			wantMatch: true,
+		},
+		{
+			name: "negative neq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "neq",
+				Value:    "42.0",
+			},
+			value: "42.0",
+		},
+		{
+			name: "lt",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "lt",
+				Value:    "42.0",
+			},
+			value:     "8",
+			wantMatch: true,
+		},
+		{
+			name: "negative lt",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "lt",
+				Value:    "42.0",
+			},
+			value: "50",
+		},
+		{
+			name: "lte",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "lte",
+				Value:    "42.0",
+			},
+			value:     "42.0",
+			wantMatch: true,
+		},
+		{
+			name: "negative lte",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "lte",
+				Value:    "42.0",
+			},
+			value: "102.0",
+		},
+		{
+			name: "gt",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "gt",
+				Value:    "10.11",
+			},
+			value:     "10.12",
+			wantMatch: true,
+		},
+		{
+			name: "negative gt",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "gt",
+				Value:    "10.11",
+			},
+			value: "1",
+		},
+		{
+			name: "gte",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "gte",
+				Value:    "10.11",
+			},
+			value:     "10.11",
+			wantMatch: true,
+		},
+		{
+			name: "negative gte",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "gte",
+				Value:    "10.11",
+			},
+			value: "0.11",
+		},
+		{
+			name: "empty value",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "0.11",
+			},
+		},
+		{
+			name: "unknown operator",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "foo",
+				Value:    "0.11",
+			},
+			value: "0.11",
+		},
+		{
+			name: "negative suffix empty value",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "suffix",
+				Value:    "bar",
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		var (
+			constraint = tt.constraint
+			value      = tt.value
+			wantMatch  = tt.wantMatch
+			wantErr    = tt.wantErr
+		)
+
+		t.Run(tt.name, func(t *testing.T) {
+			match, err := matchesNumber(constraint, value)
+
+			if wantErr {
+				require.Error(t, err)
+				var ierr errs.ErrInvalid
+				require.ErrorAs(t, err, &ierr)
+				return
+			}
+
+			require.NoError(t, err)
+			assert.Equal(t, wantMatch, match)
+		})
+	}
+}
+
+func Test_matchesBool(t *testing.T) {
+	tests := []struct {
+		name       string
+		constraint storage.EvaluationConstraint
+		value      string
+		wantMatch  bool
+		wantErr    bool
+	}{
+		{
+			name: "present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "present",
+			},
+			value:     "true",
+			wantMatch: true,
+		},
+		{
+			name: "negative present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "present",
+			},
+		},
+		{
+			name: "not present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "notpresent",
+			},
+			wantMatch: true,
+		},
+		{
+			name: "negative notpresent",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "notpresent",
+			},
+			value: "true",
+		},
+		{
+			name: "not a bool",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "true",
+			},
+			value:   "foo",
+			wantErr: true,
+		},
+		{
+			name: "is true",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "true",
+			},
+			value:     "true",
+			wantMatch: true,
+		},
+		{
+			name: "negative is true",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "true",
+			},
+			value: "false",
+		},
+		{
+			name: "is false",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "false",
+			},
+			value:     "false",
+			wantMatch: true,
+		},
+		{
+			name: "negative is false",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "false",
+			},
+			value: "true",
+		},
+		{
+			name: "empty value",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "false",
+			},
+		},
+		{
+			name: "unknown operator",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "foo",
+			},
+			value: "true",
+		},
+	}
+
+	for _, tt := range tests {
+		var (
+			constraint = tt.constraint
+			value      = tt.value
+			wantMatch  = tt.wantMatch
+			wantErr    = tt.wantErr
+		)
+
+		t.Run(tt.name, func(t *testing.T) {
+			match, err := matchesBool(constraint, value)
+
+			if wantErr {
+				require.Error(t, err)
+				var ierr errs.ErrInvalid
+				require.ErrorAs(t, err, &ierr)
+				return
+			}
+
+			require.NoError(t, err)
+			assert.Equal(t, wantMatch, match)
+		})
+	}
+}
+
+func Test_matchesDateTime(t *testing.T) {
+	tests := []struct {
+		name       string
+		constraint storage.EvaluationConstraint
+		value      string
+		wantMatch  bool
+		wantErr    bool
+	}{
+		{
+			name: "present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "present",
+			},
+			value:     "2006-01-02T15:04:05Z",
+			wantMatch: true,
+		},
+		{
+			name: "negative present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "present",
+			},
+		},
+		{
+			name: "not present",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "notpresent",
+			},
+			wantMatch: true,
+		},
+		{
+			name: "negative notpresent",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "notpresent",
+			},
+			value: "2006-01-02T15:04:05Z",
+		},
+		{
+			name: "not a datetime constraint value",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "bar",
+			},
+			value:   "2006-01-02T15:04:05Z",
+			wantErr: true,
+		},
+		{
+			name: "not a datetime context value",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value:   "foo",
+			wantErr: true,
+		},
+		{
+			name: "eq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value:     "2006-01-02T15:04:05Z",
+			wantMatch: true,
+		},
+		{
+			name: "eq date only",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "2006-01-02",
+			},
+			value:     "2006-01-02",
+			wantMatch: true,
+		},
+		{
+			name: "negative eq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value: "2007-01-02T15:04:05Z",
+		},
+		{
+			name: "neq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "neq",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value:     "2007-01-02T15:04:05Z",
+			wantMatch: true,
+		},
+		{
+			name: "negative neq",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "neq",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value: "2006-01-02T15:04:05Z",
+		},
+		{
+			name: "negative neq date only",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "neq",
+				Value:    "2006-01-02",
+			},
+			value: "2006-01-02",
+		},
+		{
+			name: "lt",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "lt",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value:     "2005-01-02T15:04:05Z",
+			wantMatch: true,
+		},
+		{
+			name: "negative lt",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "lt",
+				Value:    "2005-01-02T15:04:05Z",
+			},
+			value: "2006-01-02T15:04:05Z",
+		},
+		{
+			name: "lte",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "lte",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value:     "2006-01-02T15:04:05Z",
+			wantMatch: true,
+		},
+		{
+			name: "negative lte",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "lte",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value: "2007-01-02T15:04:05Z",
+		},
+		{
+			name: "gt",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "gt",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value:     "2007-01-02T15:04:05Z",
+			wantMatch: true,
+		},
+		{
+			name: "negative gt",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "gt",
+				Value:    "2007-01-02T15:04:05Z",
+			},
+			value: "2006-01-02T15:04:05Z",
+		},
+		{
+			name: "gte",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "gte",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value:     "2006-01-02T15:04:05Z",
+			wantMatch: true,
+		},
+		{
+			name: "negative gte",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "gte",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value: "2005-01-02T15:04:05Z",
+		},
+		{
+			name: "empty value",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "eq",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+		},
+		{
+			name: "unknown operator",
+			constraint: storage.EvaluationConstraint{
+				Property: "foo",
+				Operator: "foo",
+				Value:    "2006-01-02T15:04:05Z",
+			},
+			value: "2006-01-02T15:04:05Z",
+		},
+	}
+
+	for _, tt := range tests {
+		var (
+			constraint = tt.constraint
+			value      = tt.value
+			wantMatch  = tt.wantMatch
+			wantErr    = tt.wantErr
+		)
+
+		t.Run(tt.name, func(t *testing.T) {
+			match, err := matchesDateTime(constraint, value)
+
+			if wantErr {
+				require.Error(t, err)
+				var ierr errs.ErrInvalid
+				require.ErrorAs(t, err, &ierr)
+				return
+			}
+
+			require.NoError(t, err)
+			assert.Equal(t, wantMatch, match)
+		})
+	}
+}
