+++ internal/server/middleware/grpc/middleware.go
@@ -15,6 +15,7 @@ import (
 	"go.flipt.io/flipt/internal/server/metrics"
 	flipt "go.flipt.io/flipt/rpc/flipt"
 	fauth "go.flipt.io/flipt/rpc/flipt/auth"
+	"go.flipt.io/flipt/rpc/flipt/evaluation"
 	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
@@ -73,6 +74,28 @@ func ErrorUnaryInterceptor(ctx context.Context, req interface{}, _ *grpc.UnarySe
 // Note: this should be added before any caching interceptor to ensure the request id/response fields are unique.
 func EvaluationUnaryInterceptor(ctx context.Context, req interface{}, _ *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (resp interface{}, err error) {
 	switch r := req.(type) {
+	case *evaluation.EvaluationRequest:
+		startTime := time.Now()
+
+		// set request ID if not present
+		if r.RequestId == "" {
+			r.RequestId = uuid.Must(uuid.NewV4()).String()
+		}
+
+		resp, err = handler(ctx, req)
+		if err != nil {
+			return resp, err
+		}
+
+		// set response fields
+		if resp != nil {
+			if rr, ok := resp.(*evaluation.VariantEvaluationResponse); ok {
+				rr.Timestamp = timestamp.New(time.Now().UTC())
+				rr.RequestDurationMillis = float64(time.Since(startTime)) / float64(time.Millisecond)
+			}
+			return resp, nil
+		}
+
 	case *flipt.EvaluationRequest:
 		startTime := time.Now()
